+++ apis/python/src/tiledbsoma/_soma_group.py
@@ -18,24 +18,35 @@ from . import _tdb_handles
 # This package's pybind11 code
 from . import pytiledbsoma as clib
 from ._exception import SOMAError, is_does_not_exist_error
-from ._soma_object import AnySOMAObject, SOMAObject
-from ._types import OpenTimestamp
+from ._soma_object import SOMAObject
+from ._types import OpenTimestamp, SOMABaseTileDBType
 from ._util import is_relative_uri, make_relative_path, sanitize_key, uri_joinpath
+from .options import SOMATileDBContext
 
-CollectionElementType = TypeVar("CollectionElementType", bound=AnySOMAObject)
-_TDBO = TypeVar("_TDBO", bound=SOMAObject)  # type: ignore[type-arg]
+CollectionElementType = TypeVar("CollectionElementType", bound=SOMAObject)
+_TDBO = TypeVar("_TDBO", bound=SOMAObject)
 
 
 @attrs.define()
 class _CachedElement:
     """Item we have loaded in the cache of a collection."""
 
-    entry: _tdb_handles.GroupEntry
-    soma: AnySOMAObject | None = None
+    uri: str
+    tiledb_type: SOMABaseTileDBType | None = None
+    soma: SOMAObject | None = None
     """The reified object, if it has been opened."""
 
+    @classmethod
+    def from_handle_entry(cls, obj: tuple[str, str]) -> _CachedElement:
+        uri, type = obj[0], obj[1]
+        if type == "SOMAArray":
+            return _CachedElement(uri, SOMABaseTileDBType.SOMAArray)
+        if type == "SOMAGroup":
+            return _CachedElement(uri, SOMABaseTileDBType.SOMAGroup)
+        raise SOMAError(f"internal error: unknown object type {uri}")
 
-class SOMAGroup(SOMAObject[_tdb_handles.SOMAGroupWrapper[Any]], Generic[CollectionElementType]):
+
+class SOMAGroup(SOMAObject, Generic[CollectionElementType]):
     """Base class for all SOMAGroups: CollectionBase and MultiscaleImage.
 
     Lifecycle:
@@ -46,11 +57,16 @@ class SOMAGroup(SOMAObject[_tdb_handles.SOMAGroupWrapper[Any]], Generic[Collecti
 
     def __init__(
         self,
-        handle: _tdb_handles.SOMAGroupWrapper[Any],
+        handle: _tdb_handles.RawHandle,
+        *,
+        uri: str,
+        context: SOMATileDBContext,
         **kwargs: Any,  # noqa: ANN401
     ) -> None:
-        super().__init__(handle, **kwargs)
-        self._contents = {key: _CachedElement(entry) for key, entry in handle.initial_contents.items()}
+        super().__init__(handle, uri=uri, context=context, **kwargs)
+        self._contents = {
+            name: _CachedElement.from_handle_entry(entry) for name, entry in self._handle.members().items()
+        }
         """The contents of the persisted TileDB Group.
 
         This is loaded at startup when we have a read handle.
@@ -75,14 +91,13 @@ class SOMAGroup(SOMAObject[_tdb_handles.SOMAGroupWrapper[Any]], Generic[Collecti
             if entry.soma is None:
                 from . import _factory  # Delayed binding to resolve circular import.
 
-                uri = entry.entry.uri
-                mode = self.mode
-                context = self.context
-                timestamp = self.tiledb_timestamp_ms
-                clib_type = entry.entry.wrapper_type.clib_type
-
-                wrapper = _tdb_handles.open_handle_wrapper(uri, mode, context, timestamp, clib_type)
-                entry.soma = _factory.reify_handle(wrapper)
+                entry.soma = _factory._open_soma_object(
+                    uri=entry.uri,
+                    mode=self.mode,
+                    context=self.context,
+                    tiledb_timestamp=self.tiledb_timestamp_ms,
+                    clib_type=None if entry.tiledb_type is None else entry.tiledb_type.name,
+                )
 
                 # Since we just opened this object, we own it and should close it.
                 self._close_stack.enter_context(entry.soma)
@@ -113,7 +128,7 @@ class SOMAGroup(SOMAObject[_tdb_handles.SOMAGroupWrapper[Any]], Generic[Collecti
             obj = entry.soma
             if obj is None:
                 # We haven't reified this SOMA object yet. Don't try to open it.
-                yield f"{indent}{key!r}: {entry.entry.uri!r} (unopened)"
+                yield f"{indent}{key!r}: {entry.uri!r} (unopened)"
             else:
                 yield f"{indent}{key!r}: {obj._my_repr()}"
                 if isinstance(obj, SOMAGroup):
@@ -151,10 +166,7 @@ class SOMAGroup(SOMAObject[_tdb_handles.SOMAGroupWrapper[Any]], Generic[Collecti
             name=key,
             soma_type=soma_object.soma_type,
         )
-        self._contents[key] = _CachedElement(
-            entry=_tdb_handles.GroupEntry(soma_object.uri, soma_object._wrapper_type),
-            soma=soma_object,
-        )
+        self._contents[key] = _CachedElement(uri=soma_object.uri, tiledb_type=None, soma=soma_object)
         self._mutated_keys.add(key)
 
     def _del_element(self, key: str) -> None:
@@ -269,7 +281,9 @@ class SOMAGroup(SOMAObject[_tdb_handles.SOMAGroupWrapper[Any]], Generic[Collecti
             Experimental.
         """
         super().reopen(mode, tiledb_timestamp)
-        self._contents = {key: _CachedElement(entry) for key, entry in self._handle_wrapper.initial_contents.items()}  # type: ignore[union-attr]
+        self._contents = {
+            name: _CachedElement.from_handle_entry(entry) for name, entry in self._handle.members().items()
+        }
         self._mutated_keys = set()
         return self